fix output handling in xcatclient vswitch query iuo stats
In xcat response data, vsw query iuo stats are splited to a number
of parts.

It is expected vsw query iuo stats data in several sections for
each vswitch, and a blank line between each vswitch section.
However, in some special cases, the xcat response data separated
at the end of a vswitch section, will cause the blank line lost.

Fix this by insert a blank line in this case.

Change-Id: I37ce505e6153b23b8e7fda2e3ba6ff4a6c781cfd