Update patch set 1
Patch Set 1: Code-Review-1

Overall, I suggest you put out the invokeSMCLI created message and then put out a message that refines the information.  This allows the customer to see the specific cause and more information. It also lessens the need for the routines that use invokeSMCLI to need to know the invocation format, location, SMCLI, --addcheader.
I suggest you consider whether there are times when you want to update the rc and rs based on your refinement of the error.

Patch-set: 1
Label: Code-Review=-1