Update patch set 2
Patch Set 2: -Code-Review

I looked for the other conversion routines and I think we could actually use the cvtToMag routine in generalUtils.py.  And yes, based on other examples, we don't need to add in any extra error checking.  Sorry I was thinking we had added that, but we decided if the SMAPI return codes were okay then we didn't need to.

Patch-set: 2
Label: Code-Review=0