Solve the iucv install and upgrade bug in ubuntu
The submit contains the changes:
Using /bin/bash instead of /bin/sh otherwise it will not work on
ubuntu since in these scripts the function keyword is a bashism,
a bash extension. POSIX syntax does not use function.

Print the log of iucvclnt to where the syslog print to keep same
as iucvserver log.

Modify some log output to console.

Change-Id: Ibaf7223ea0b16bc9424ae4ef9bc333bbd56eefdc