Database enhancement of connection pools
It's better to make volumeop, networkop, or vmop each use one
seperate database connection object, in case chaos. For example,
we may need call rollback() function in case of failure in the
future, another one's operation can be rollback if all use a same
connection object.
Also I provide a sample of how to get a database connection and
execute SQL statements in volumop.VolumeUtils, please refer it
to write your code. It's a temporary sample so it may disapper
in short time, but not shorter than some practical samples are
there.
In the end, feel free to modify and improve it!

Change-Id: Ie3311c4e143705d3e65d3cb8fe62eedcd647f351