Update patch set 1
Patch Set 1:

Question on the patch, is the RC=0 that we're testing for from the command getting sent to the guest, or will that also include that the command was successful?  I just wondered if we need some additional output verification added to make sure the online of the device worked?

Patch-set: 1