Update patch set 1
Patch Set 1:

overallRC = 0 indicates that a command was "successfully" issued on the managed system.  The failure case that I am fixing resulted in this value being non-zero because the chccwdev failed due to the fact that the device subsystem had not completed other work and responded to the chccwdev that it could not handle the command.  The settle causes the device subsystem to complete its other work before we submit the chccwdev to enable the device.  The code went from failing every 2nd or 4th time on my system to not failing in 19 runs.

Patch-set: 1