Update patch set 4
Patch Set 4:

I tested the option. If leave it default as True, will got following error when using the same db_conn:  
ProgrammingError: SQLite objects created in a thread can only be used in that same thread.The object was created in thread id 139997841147712 and this is thread id 139997690853120  

And I also tested using db connection with check_same_thread=False, it worked fine.  

So the results is: if we want to init all the shared db connection at the beginning, should set check_same_thread=False, in order to be used in sub-thread.

Patch-set: 4