Update patch set 1
Patch Set 1: Code-Review+1

I am ok with the fix for smutTest.  Does the SDK do a similar function so that it will not encounter the problem?

I talked to Patty Rando about the problem.  She did not have a clear idea why your solution addresses the problem.  She said that if she had a log from the time of the problem that she would have a better idea on the cause and whether this solution is the best.  You may want to consider that as a follow-up.

Patch-set: 1