Update patch set 1
Patch Set 1:

I talked with Scott Coyle.  He and I do not think that the use of Static_Image_Changes_Immediate_DM should NOT be considered.  The documentation states that DirMaint immediately invokes directxa to create the object directory BUT does not ensure that all tasks in the queue to update the source directory have been completed.  This has the potential to cause other problems.

I tried recreating the failure on my system to see the log file from the smutTest run but could not cause the problem.  I think it would be helpful if you provided at least the output of the smutTest run.  Both Scott and Patty think that the logs from SMAPI and DirMaint should be reviewed to get a better understanding of the underlying issue.

The ultimate cause of the problem appears to be a timing issue where the SMUT layer is too quick to add a disk after creating the system.  The Image_Query_DM is probably just allowing time to elapse between creating the system is and adding disk.  Your previous reply did not state why a user of the SDK would not encounter the same problem that smutTest encountered.  I would like to know why the SDK does not need a change or what in the SDK will prevent this problem.

Patch-set: 1