Update patch set 1
Patch Set 1: Code-Review-1

Thanks for the link to the failed test run.  After reviewing it, I have a better understanding of the issue.

Here is my summary of the issue.  The smcli Image_Create_DM returns rc = 0 after the directory manager has added the userid to the source directory but possibly before it has had a chance to update the object directory.  This occurs in smutTest rather than an xCAT environment because the tests code flow through smutTest does less before attempting to use the virtual machine by tranferring a reader file to it.  This problem could exist for xCAT but we do not normally see it because we delay enough due to other xCAT processing so that the object directory is updated before the next command that needs the user to exist.

The SDK could encounter a similar issue with SMUT.  Thus, a change to smutTest should be considered only a short term patch.  The recommended fix would be to change the makeVM function to support an option that says "don't return until the virtual machine exists".  I suggest that this be a set of options similar to what is provided on the powerVM functions: [ --wait ] [ --maxwait secs ] [ --poll secs ].  A suitable default could be used when on --wait is specified (i.e. --maxwait 30 and --poll 10).  I recommend --wait as an option as opposed to a default action because there are times when you are doing multiple directory changes to a system so that I would expect to not need --wait.  This is because in that case we are working with the source directory through the directory manager.

A change in makevm processing could optionally loop using the poll interval and issue a "vmcp q <userid>".  When it no longer gets a 003E message (HCPCQV003E Invalid option - ST63D1) or it gets a 045E message (HCPCQU045E ST63D1L not logged on) then it knows the object directory has been updated.  On my system, it is almost immediate but I suspect on your system it may be taking a few seconds.  

The other more expensive alternative would be to implement an event listener.  I do not recommend that for this problem at this time. 

The short term patch to smutTest to hide this issue would be to use a SHELL invocation.  You can do this with the following two changes.

To the localSubs dictionary add the following line:
'<<<makevmWait>>>': '15',                   # Wait time for makeVM to fully complete

After the makevm call in the modify test set add the following lines instead of your proposed change:
    {
        'description': "Wait <<<makevmWait>>> seconds for source " +
            "directory to be updated.",
        'request': "SHELL echo 'Sleeping for <<<makevmWait>>> seconds " +
            "to allow source directory update to complete';sleep " +
            "<<<makevmWait>>>",
        'out': "",
        'overallRC': [0],
    },

Patch-set: 1
Label: Code-Review=-1