Add return code definition rules and simple example
All the return code needs to be added into the errors list in
returncode.py. The definiton rules are specified in the start
part of the file.

And when raise a exception, we need to make sure all exceptions
are defined with a results attribute.

Rewrite guest_create and guest_deploy as the usage example.

Change-Id: Iaa2f5ec0b1fa57bbcc420beb90b5996c17995ce5