Update patch set 3
Patch Set 3:

In reply to YiChun's comments:

As for rc, actually it is already defined in ModRCs in returncode.py, you can use (actually only exception.py would use) it by ModRCs['guest'], ModRCs['volume'], etc.

As for overallRC, it has only a few value. All the errors are grouped by the overallRC value as defined in the dict 'errors' in returncode.py. As long as it is defined in the errorcode.py, the upper layer (xxops, smutclient) needs to care about the value. The errorcode.py mostly corperate with exception.py to define the exceptions, and upper layer just need to care about the exceptions.

Patch-set: 3
Reviewer: DongYanYang <1008926@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0