Actively enable network interface configuration
1. Add new paramaetr to guest db to mark whether the guest has already
set the network. it will be 0 when the guest is created. Each time adding
the new network interface, it will check this paramater. If it is 0, means
it is the first time to set interface for the guest, it will help to
remove the old network configuretion from image
2. A new file activeIPaddr is included in doscript, it will help to
make the new IP address taking effect
3. Using vdev to create interface name instead of device_num, it is
required by active enablement, device_num will overwrite the previous
network interface name

Change-Id: I1a753e0ee50741728b832590a6cb04882f3523e4