Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > -1 for hold this

 > (1 comment)
 > 
 > -1 for hold this

First: it is used to define whether this is the first time for a user to create network interface for the guest. IF it is the first time, set it to be True, this will make the 

 > (1 comment)
 > 
 > -1 for hold this

First: it is used to define whether this is the first time that the user creates the network interface for the guest. It is introduced because the image may be captured from a guest, and it will include the network interfaces files of the guest. When the user uses this image to deploy a new guest, these network interface files will be copied to the new guest. So for the first time that the user creates his own network interfaces on the guest, the old files need to be removed. The parameter 'first' will tell the operator whether to remove the network files. If it is the first time, it should be true, to make sure the old network interfaces files removed. If it is not, it must be False, otherwise it will erase the current interface network configuration

Patch-set: 2
Reviewer: ylpan@cn.ibm.com <1010999@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0