checkout arguments format
fix bug 1744222
when calling REST APIs, if some arguments should be put into
args transferred in as kwargs, the loop will raise exception
because out of index, by contrast, kwargs members transferred as
args will cause problems of missing data.
this patch check the amount of args, if the amount is not equal to
the required numbers, will raise ArgsFormatError.

Change-Id: I0e256da9ed9433665f2a41c5aecfb6581c62df05