change default value of token_path to ''
Before this patch, user need to generate a token in default token path
whether the WSGI will validate the token or not.
Now change the default value to token_path to '' and if user does not
specify a token file, the sdkserver will raise an error of token
incorrect unless configuring the way of auth to none on the server side.

Change-Id: I063e76c01d1e0967d25e51f932ad5c3983ae7c72