adding multipath volumes to same userid
now adding more than one volumes to same userid will
meet an error because my old understanding is wrong.
The correct though is, when we attaching multipul volumes to same
userid we should use the FCP device it already own. If we attach
volume to a another userid, we use a new FCP device.

Signed-off-by: SharpRazor <bjcb@cn.ibm.com>