Write DNS info to network conf file for rhel
This commit is related to issue:
https://github.ibm.com/zvc/planning/issues/591

This issues reports a problem, if DNS nameserver info
is specified when creating a network in icic, we expect
it takes effect inside VMs created using this network.
While currently DNS nameserver info is not written to
/etc/resolv.conf.

There are 2 reasons causing problem above:
1. nova zvm driver doesn't read DNS info from openstack
   subnet.
2. /etc/resolv.conf probably is overwritten by NetworkManager
   due to the latter's working behavior. Need find a way
   align with NetworkManager to set the DNS nameserver.

This commit is to solve the 2nd problem. For Rhel,
NetworkManger can read DNS1, DNS2 value from
/etc/sysconfig/network-scripts/ifcfg-enccw0.0.1000 file
and write to /etc/resolv.conf
So this commit is to write DNS nameserver info into
/etc/sysconfig/network-scripts/ifcfg-enccw0.0.1000 as DNS1, DNS2
if DNS nameserver info is configured.

Signed-off-by: Shu Juan Zhang <zshujuan@cn.ibm.com>