fix bug of adding multi volumes
When attaching the volumes, the UI shows success, but 'multipath -l'
show nothing about the volumes coming after first one.
This is because we did not process all the WWPNs when confiurating the
zlinux.
What's more, Our zlinux discovery the WWPNs itself because target_wwpns
given by openstack contains inative WWPNs.
So we choose to dicovery the WWPNs in the OS and do not use the target_wwpns
given by openstack.

Signed-off-by: SharpRazor <bjcb@cn.ibm.com>