update guest_deploy to support boot from volume
add a new parameter skipdiskcopy as when boot_from_volume we won't rely
on unpackdiskimage to do disk copy.
when skipdiskcopy is False, image_name is the name of image as original.
when True, image_name will be used as the os_version which will be
used to update the guest database table.

Signed-off-by: dyyang <dyyang@cn.ibm.com>