Limit the length of rd.zfcp when refresh bootmap
In the boot from volume scenario for RHEL, the VM may fail if the length of kernel parameter
line exceeds the maximum allowed. So the production recommend is 8 different paths of rd.zfcp.

If the count of FCP devices is greater than 8, it will report an error and exit directly.

If the count of WWPN is greater than 8, it will report a warning and don't exit.

Signed-off-by: changzhi <changzhi1990@gmail.com>