update refreshbootmap API to return a dict
Previously we return the FCPs and WWPNs in a seperate list
which assumes all the FCP devices have access to all the WWPNs.
This does not stand for all customer scenarios.
This change will find all valid paths while attaching the volume.
and return the valid paths in a dict of format:
{'FCP1': ['W1','W2'], "FCP2": ['W3', 'W4']}

Signed-off-by: dyyang <dyyang@cn.ibm.com>