don't mount the /run folder for chroot zipl
we have find that the mount of /run folder in a chroot case would
cause the system down, (the /run folder files in the system is lost.)

and I checked several pages related to chroot without the mount of /run folder
and some page explicitly mentioned /run is not recommended to be mounted.

so remove the mount of /run from our code.

Signed-off-by: dyyang <dyyang@cn.ibm.com>