Add interfaces to get disk volumes of the diskpool
As for CIC onboarding feature, we need check the diskpool of
one vm is the same as the one configued for the sdkserver,
for which we can match the volumes with the one in the vm user direct.
So add these interfaces to get the disk volumes of the diskpool.

If the diskpool is provided, query the voluems of the pool, if not,
the CONF.zvm.disk_pool will be used.

Signed-off-by: Xiao Feng Ren <renxiaof@linux.vnet.ibm.com>