[master] zipl without chroot
In the current refreshbootmap implementation, we use chroot to do zipl to refresh the bootloader of guest vm. In order to be able to chroot, we need to bind mount several system dirs of the compute node

This implementation has been proofed to sometimes cause the compute system down： #2427

It is dangerous to mount system dirs and chroot. And zipl command has the support of specifying the initramfs/kernel path parameter to refresh bootloader.

With this PR:
- it changes temporary mount point from /mnt to /tmp
- it enable zipl for RHEL7/RHEL8 without chroot by specifing the path of zipl.conf and BootLoaderSpec dir.
```shell

  ${deviceMountPoint}/usr/sbin/zipl -V \
    -c ${deviceMountPoint}/etc/zipl.conf \
    -b ${deviceMountPoint}/boot/loader/entries
  ${deviceMountPoint}/usr/sbin/zipl -V \
    -c ${deviceMountPoint}/etc/zipl.conf
```

Signed-off-by: Da Long Wang <shdlwang@cn.ibm.com>