Add check if user directory exists in guest_get_power_state
Currently guest_get_power_state returns shutdown if the guest userid
exists in DB but does not exist in user directory (e.g. it has been
deleted through smcli directly), and this is wrong which could cause
confusion and unexpected result.
Also because if the userid does not exist in DB, it returns 404.
Hence, the fix returns 404 to caller in this case.

Signed-off-by: QingFeng Hao <haoqf@linux.vnet.ibm.com>