fix the loop device lost when reboot
with PR: https://github.com/openmainframeproject/feilong/pull/336/files
the loop device will only be setup when the cfgdrive is first recieved.
But the loop device will be lost at system reboot which will cause cloud-init
to think the first reboot is the first boot.

This PR will check whether the device with config-2 label (which cloud-init use as configdrive souce)
 already exists, create the loop device only if it does not exist.

Signed-off-by: dyyang <dyyang@cn.ibm.com>