try to deactivate lv and flush mpath again before exit the refreshbootmap
In some case, though the `multipath -f` reported the multipath device is flushed successfully, multipathd will continue to add new path
into the mpath device making the mpath device showing up again and the LVs on it then automatically be activated.
So we add the check of mpath existence again at the end and do the deactivate of LVs and flush mpath device, to try our best to avoid leftover issue.

Signed-off-by: dyyang <dyyang@cn.ibm.com>