add wwid as a parameter for refresh_bootmap
Before this change, the wwid is get from the scsi_id cmd on a single path after we find
a valid disk show up on the compute node.
But in some error case that the refreshbootmap may fail in very early stage before we find a valid path, eg
when failed to online some FCPs, or no valid path found on any FCPs, in these cases, the wwid maybe "". Then
when doing cleanup at the exit of refreshbootmap, we cann't get the multipath map name (eg, mpatha) from wwid and
then we cann't do the cleanup of the mpath and disk.

so this change would accept wwid as a parameter and use it directly in the cleanup.

Signed-off-by: dyyang <dyyang@cn.ibm.com>