update rhel7 kernel cmdline to use mpath as the root
previously we only hardcode to mpath path as root for "RHEL8".
For RHEL7, we still use the default root fs UUID. But in our test, we
found that RHEL7 may also use single path as root when system boot.
This commit change the RHEL7 to explicitly specify mpath path as root.

Signed-off-by: dyyang <dyyang@cn.ibm.com>