enhance to leftover handling
previously we will check the mpath existence again after the lun removed
and fcp devices offline, because we have found there's chance that the multipathd
readded the map again after it is flushed.

But the check will based on whether the /dev/mapper/$map_name file exists. Recently
we find that there's scenario the /dev/mapper/$map_name does not exist while it exists
in the output of 'multipath -l'.

So in this fix, we removed the check of '/dev/mapper/$map_name'， just do the leftover
check and flush again at the end, no side-effect of doing this.

Signed-off-by: dyyang <dyyang@cn.ibm.com>