Remove the FCP database rollback commited in PR: https://github.com/openmainframeproject/feilong/pull/538
In previous PR: https://github.com/openmainframeproject/feilong/pull/538
we have this:

> To make the FCP database consistent with the FCPs defined to vm, compare the valid_path_string
>returned by refreshbootmap, if some fcps allocated but found not usable to be defined in vm, then
>mark the fcp unreserved in database.

But this would cause the FCP marked as reserved=0 and connections=1, and also since these FCPs are
already defined in the storage host, we cann't mark them as free.

So this patch would remove the check and keep the unusable FCPs marked as in-use.

Signed-off-by: dyyang <dyyang@cn.ibm.com>