Skip replace direct and nic couple when vswitch exists in NICDEF
In couple_nic_to_vswitch, Image Directory is replaced by adding
vswitch to NICDEF statement. If the user directory NICDEF
statment already included the specified switch, means the NIC has
already coupled to the vswitch, error happens when running above
actions.

This commit is to skip replace direct and nic couple when
 vswitch exists in NICDEF. A user scenario is when a VM is migrating
to a target host. The target host detected the device add and call
into couple nic to vswitch.

Signed-off-by: Shu Juan Zhang <zshujuan@cn.ibm.com>