Support resize memory when USER Directory Statement has more than 6 fields
Current implementation in smtclient.py, if USER Directory Statement has
fields number not equal to 6, then defined memory and reserved memory value
of the userid are fetched as -1, which leads to memory resize failure.

With this commit, it is not strictly requires the USER Directory Statement
have 6 fields, because there are real cases that has 7 fields.
Now it is to check the fields number is not fewer than 6.

Signed-off-by: Shu Juan Zhang <zshujuan@cn.ibm.com>